refactor: share validation
