+++ packages/content/src/attributes/types/identity/SchematizedJSON.ts
@@ -1,5 +1,7 @@
 import { serialize, type, validate } from "@js-soft/ts-serval";
 import { AbstractAttributeValue, AbstractAttributeValueJSON, IAbstractAttributeValue } from "../../AbstractAttributeValue";
+import { RenderHints, RenderHintsEditType, RenderHintsTechnicalType, ValueHints } from "../../hints";
+import { validateJSON } from "../utils";
 
 export interface SchematizedJSONJSON extends AbstractAttributeValueJSON {
     "@type": "SchematizedJSON";
@@ -15,7 +17,7 @@ export interface ISchematizedJSON extends IAbstractAttributeValue {
 @type("SchematizedJSON")
 export class SchematizedJSON extends AbstractAttributeValue implements ISchematizedJSON {
     @serialize({ any: true })
-    @validate({ customValidator: validateValue })
+    @validate({ customValidator: validateJSON })
     public value: unknown;
 
     @serialize()
@@ -29,6 +31,17 @@ export class SchematizedJSON extends AbstractAttributeValue implements ISchemati
     })
     public schemaURL?: string;
 
+    public static get valueHints(): ValueHints {
+        return ValueHints.from({});
+    }
+
+    public static get renderHints(): RenderHints {
+        return RenderHints.from({
+            editType: RenderHintsEditType.TextArea,
+            technicalType: RenderHintsTechnicalType.Unknown
+        });
+    }
+
     public static from(value: ISchematizedJSON | Omit<SchematizedJSONJSON, "@type">): SchematizedJSON {
         return this.fromAny(value);
     }
@@ -36,21 +49,4 @@ export class SchematizedJSON extends AbstractAttributeValue implements ISchemati
     public override toJSON(verbose?: boolean | undefined, serializeAsString?: boolean | undefined): SchematizedJSONJSON {
         return super.toJSON(verbose, serializeAsString) as SchematizedJSONJSON;
     }
-}
-
-function validateValue(value: any) {
-    try {
-        const string = JSON.stringify(value);
-        if (string.length > 4096) {
-            return "stringified value must not be longer than 4096 characters";
-        }
-    } catch (e) {
-        if (e instanceof SyntaxError) {
-            return "must be a valid JSON object";
-        }
-
-        return "could not validate value";
-    }
-
-    return undefined;
 }
