+++ packages/content/src/attributes/types/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./validateJSON";
